requirements.txt
Flask
Flask-RESTful
py-zabbix 
